+++ src/Symfony/Component/Messenger/CHANGELOG.md
@@ -7,7 +7,7 @@ CHANGELOG
  * The component is not experimental anymore
  * All the changes below are BC BREAKS
  * `MessageBusInterface::dispatch()` and `MiddlewareInterface::handle()` now return `void`
- * `MiddlewareInterface::handle()` now require an `Envelope` as first argument
+ * `MiddlewareInterface::handle()` now require an `Envelope` as first argument and a `StackInterface` as second
  * `EnvelopeAwareInterface` has been removed
  * The signature of `Amqp*` classes changed to take a `Connection` as a first argument and an optional
    `Serializer` as a second argument.
@@ -16,7 +16,6 @@ CHANGELOG
    Instead, it accepts the sender instance itself instead of its identifier in the container.
  * `MessageSubscriberInterface::getHandledMessages()` return value has changed. The value of an array item
    needs to be an associative array or the method name.
- * `ValidationMiddleware::handle()` and `SendMessageMiddleware::handle()` now require an `Envelope` object
  * `StampInterface` replaces `EnvelopeItemInterface` and doesn't extend `Serializable` anymore
  * The `ConsumeMessagesCommand` class now takes an instance of `Psr\Container\ContainerInterface`
    as first constructor argument
@@ -30,9 +29,16 @@ CHANGELOG
  * `SenderLocatorInterface::getSenderForMessage()` has been replaced by `getSender(Envelope $envelope)`
  * `MessengerDataCollector::getMessages()` returns an iterable, not just an array anymore
  * `AbstractHandlerLocator` is now internal
- * `HandlerLocatorInterface::resolve()` has been replaced by `getHandler(Envelope $envelope)`
+ * `HandlerLocatorInterface::resolve()` has been replaced by `getHandler(Envelope $envelope): ?callable` and shouldn't throw when no handlers are found
  * `SenderLocatorInterface::getSenderForMessage()` has been replaced by `getSender(Envelope $envelope)`
  * `SenderInterface::send()` returns `void`
+ * Classes in the `Middleware\Enhancers` sub-namespace have been moved to the `Middleware` one
+ * Classes in the `Asynchronous\Routing` sub-namespace have been moved to the `Transport\Sender\Locator` sub-namespace
+ * The `Asynchronous/Middleware/SendMessageMiddleware` class has been moved to the `Middleware` namespace
+ * `SenderInterface` and `ChainSender` classes have been moved to the `Transport\Sender` sub-namespace
+ * `ReceiverInterface` and its implementations have been moved to the `Transport\Receiver` sub-namespace
+ * `ActivationMiddlewareDecorator` has been renamed `ActivationMiddleware`
+ * `AllowNoHandlerMiddleware` has been removed in favor of a new constructor argument on `HandleMessageMiddleware`
 
 4.1.0
 -----